@@ -1,9 +1,10 @@
 name: example
 description: Example project.
 version: 0.0.1
+publish_to: none
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   dog:
