@@ -1,14 +1,14 @@
 name: example_web
 description: Use dog in dart web.
+publish_to: none
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   dog:
     path: ../
 
 dev_dependencies:
-  build_runner: ^1.10.0
-  build_web_compilers: ^2.11.0
-  pedantic: ^1.9.0
+  build_runner: ">=2.0.0 <3.0.0"
+  build_web_compilers: ">=3.0.0 <4.0.0"
