@@ -5,13 +5,14 @@ repository: https://github.com/chenenyu/dog
 publish_to: https://pub.dev
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  ansicolor: ^1.1.0
-  stack_trace: ^1.9.5
-  tuple: ^1.0.3
+  ansicolor: ^2.0.1
+  stack_trace: ^1.10.0
+  tuple: ^2.0.0
 
 dev_dependencies:
-  http: ^0.12.2
-  path: ^1.7.0
+  http: ^0.13.0
+  path: ^1.8.0
+  lints: ^1.0.0
