@@ -1,8 +1,9 @@
 import 'dart:convert';
 import 'dart:math';
 
-import '../formatter.dart';
-import '../record.dart';
+import 'package:dog/src/formatter.dart';
+import 'package:dog/src/record.dart';
+import 'package:dog/src/util/utils.dart';
 
 /// This formatter will not generate borders.
 class SimpleFormatter extends Formatter {
@@ -10,31 +11,31 @@ class SimpleFormatter extends Formatter {
   final int stackTraceLevel;
 
   /// Function to get caller info.
-  final StringCallback callerGetter;
+  final MessageCallback? callerGetter;
 
   SimpleFormatter({
     this.stackTraceLevel = 10,
-    this.callerGetter = callerInfo,
+    this.callerGetter = DogUtils.defaultCallerInfo,
   });
 
   @override
   List<String> format(Record record) {
     List<String> lines = [];
     // tag/level time caller
-    String caller = callerGetter == null ? null : callerGetter();
+    String? caller = callerGetter?.call().toString();
     lines.add('${record.tag ?? record.level.name}'
         ' ${record.dateTime.toIso8601String()}'
         '${caller == null ? '' : (' (' + caller + ')')}');
     // title
     if (record.title != null) {
-      lines.add(record.title);
+      lines.add(record.title!);
     }
     // message
     String msg = convertMessage(record.message);
     lines.add(msg);
     // stack trace
     if (record.stackTrace != null) {
-      String st = convertStackTrace(record.stackTrace);
+      String st = convertStackTrace(record.stackTrace!);
       lines.add(st);
     }
     return lines;
@@ -51,7 +52,7 @@ class SimpleFormatter extends Formatter {
         // JsonUnsupportedObjectError
         msg = message.toString();
       }
-    } else if (message is StringCallback) {
+    } else if (message is MessageCallback) {
       msg = message().toString();
     } else if (message is Exception) {
       msg = message.toString();
