@@ -4,7 +4,8 @@ export 'src/dog.dart';
 export 'src/emitter.dart';
 export 'src/emitter/console_emitter.dart';
 export 'src/emitter/file_emitter.dart';
-export 'src/formatter.dart' hide callerInfo;
+export 'src/formatter.dart';
 export 'src/formatter/pretty_formatter.dart';
 export 'src/formatter/simple_formatter.dart';
+export 'src/handler.dart';
 export 'src/level.dart';
