@@ -1,5 +1,5 @@
 /// If you define your own level, make sure you use a value
-/// between those used in [Level.ALL] and [Level.OFF].
+/// between those used in [Level.all] and [Level.off].
 class Level implements Comparable<Level> {
   final String name;
 
@@ -10,29 +10,29 @@ class Level implements Comparable<Level> {
   const Level(this.name, this.value);
 
   /// Special key to turn on logging for all levels ([value] = 0).
-  static const Level ALL = Level('ALL', 0);
+  static const Level all = Level('all', 0);
 
   /// Special key to turn off all logging ([value] = 2000).
-  static const Level OFF = Level('OFF', 2000);
+  static const Level off = Level('off', 2000);
 
-  static const Level VERBOSE = Level('VERBOSE', 300);
+  static const Level verbose = Level('verbose', 300);
 
-  static const Level DEBUG = Level('DEBUG', 500);
+  static const Level debug = Level('debug', 500);
 
-  static const Level INFO = Level('INFO', 800);
+  static const Level info = Level('info', 800);
 
-  static const Level WARNING = Level('WARNING', 900);
+  static const Level warning = Level('warning', 900);
 
-  static const Level ERROR = Level('ERROR', 1000);
+  static const Level error = Level('error', 1000);
 
-  static const List<Level> LEVELS = [
-    ALL,
-    VERBOSE,
-    DEBUG,
-    INFO,
-    WARNING,
-    ERROR,
-    OFF
+  static const List<Level> levels = [
+    all,
+    verbose,
+    debug,
+    info,
+    warning,
+    error,
+    off,
   ];
 
   @override
