@@ -1,4 +1,4 @@
-// Generated on 2020-11-20 08:14:22.608341Z
+// Generated on 2021-10-19 09:26:46.938756Z
 part of dog.util;
 
 List<Tuple2> _ascii_table = [
@@ -19,6 +19,7 @@ List<Tuple2> _nonprint_table = [
   Tuple2(0x61C, 0x61C),
   Tuple2(0x6DD, 0x6DD),
   Tuple2(0x70F, 0x70F),
+  Tuple2(0x890, 0x891),
   Tuple2(0x8E2, 0x8E2),
   Tuple2(0x180E, 0x180E),
   Tuple2(0x200B, 0x200F),
@@ -62,7 +63,8 @@ List<Tuple2> _combining_table = [
   Tuple2(0x825, 0x827),
   Tuple2(0x829, 0x82D),
   Tuple2(0x859, 0x85B),
-  Tuple2(0x8D3, 0x8E1),
+  Tuple2(0x898, 0x89F),
+  Tuple2(0x8CA, 0x8E1),
   Tuple2(0x8E3, 0x903),
   Tuple2(0x93A, 0x93C),
   Tuple2(0x93E, 0x94F),
@@ -104,6 +106,7 @@ List<Tuple2> _combining_table = [
   Tuple2(0xBCA, 0xBCD),
   Tuple2(0xBD7, 0xBD7),
   Tuple2(0xC00, 0xC04),
+  Tuple2(0xC3C, 0xC3C),
   Tuple2(0xC3E, 0xC44),
   Tuple2(0xC46, 0xC48),
   Tuple2(0xC4A, 0xC4D),
@@ -155,13 +158,14 @@ List<Tuple2> _combining_table = [
   Tuple2(0x108F, 0x108F),
   Tuple2(0x109A, 0x109D),
   Tuple2(0x135D, 0x135F),
-  Tuple2(0x1712, 0x1714),
+  Tuple2(0x1712, 0x1715),
   Tuple2(0x1732, 0x1734),
   Tuple2(0x1752, 0x1753),
   Tuple2(0x1772, 0x1773),
   Tuple2(0x17B4, 0x17D3),
   Tuple2(0x17DD, 0x17DD),
   Tuple2(0x180B, 0x180D),
+  Tuple2(0x180F, 0x180F),
   Tuple2(0x1885, 0x1886),
   Tuple2(0x18A9, 0x18A9),
   Tuple2(0x1920, 0x192B),
@@ -170,7 +174,7 @@ List<Tuple2> _combining_table = [
   Tuple2(0x1A55, 0x1A5E),
   Tuple2(0x1A60, 0x1A7C),
   Tuple2(0x1A7F, 0x1A7F),
-  Tuple2(0x1AB0, 0x1AC0),
+  Tuple2(0x1AB0, 0x1ACE),
   Tuple2(0x1B00, 0x1B04),
   Tuple2(0x1B34, 0x1B44),
   Tuple2(0x1B6B, 0x1B73),
@@ -183,8 +187,7 @@ List<Tuple2> _combining_table = [
   Tuple2(0x1CED, 0x1CED),
   Tuple2(0x1CF4, 0x1CF4),
   Tuple2(0x1CF7, 0x1CF9),
-  Tuple2(0x1DC0, 0x1DF9),
-  Tuple2(0x1DFB, 0x1DFF),
+  Tuple2(0x1DC0, 0x1DFF),
   Tuple2(0x20D0, 0x20F0),
   Tuple2(0x2CEF, 0x2CF1),
   Tuple2(0x2D7F, 0x2D7F),
@@ -237,10 +240,14 @@ List<Tuple2> _combining_table = [
   Tuple2(0x10D24, 0x10D27),
   Tuple2(0x10EAB, 0x10EAC),
   Tuple2(0x10F46, 0x10F50),
+  Tuple2(0x10F82, 0x10F85),
   Tuple2(0x11000, 0x11002),
   Tuple2(0x11038, 0x11046),
+  Tuple2(0x11070, 0x11070),
+  Tuple2(0x11073, 0x11074),
   Tuple2(0x1107F, 0x11082),
   Tuple2(0x110B0, 0x110BA),
+  Tuple2(0x110C2, 0x110C2),
   Tuple2(0x11100, 0x11102),
   Tuple2(0x11127, 0x11134),
   Tuple2(0x11145, 0x11146),
@@ -306,6 +313,8 @@ List<Tuple2> _combining_table = [
   Tuple2(0x16FE4, 0x16FE4),
   Tuple2(0x16FF0, 0x16FF1),
   Tuple2(0x1BC9D, 0x1BC9E),
+  Tuple2(0x1CF00, 0x1CF2D),
+  Tuple2(0x1CF30, 0x1CF46),
   Tuple2(0x1D165, 0x1D169),
   Tuple2(0x1D16D, 0x1D172),
   Tuple2(0x1D17B, 0x1D182),
@@ -324,6 +333,7 @@ List<Tuple2> _combining_table = [
   Tuple2(0x1E023, 0x1E024),
   Tuple2(0x1E026, 0x1E02A),
   Tuple2(0x1E130, 0x1E136),
+  Tuple2(0x1E2AE, 0x1E2AE),
   Tuple2(0x1E2EC, 0x1E2EF),
   Tuple2(0x1E8D0, 0x1E8D6),
   Tuple2(0x1E944, 0x1E94A),
@@ -395,36 +405,62 @@ List<Tuple2> _doublewide_table = [
   Tuple2(0x17000, 0x187F7),
   Tuple2(0x18800, 0x18CD5),
   Tuple2(0x18D00, 0x18D08),
-  Tuple2(0x1B000, 0x1B11E),
+  Tuple2(0x1AFF0, 0x1AFF3),
+  Tuple2(0x1AFF5, 0x1AFFB),
+  Tuple2(0x1AFFD, 0x1AFFE),
+  Tuple2(0x1B000, 0x1B122),
   Tuple2(0x1B150, 0x1B152),
   Tuple2(0x1B164, 0x1B167),
   Tuple2(0x1B170, 0x1B2FB),
-  Tuple2(0x1F200, 0x1F200),
-  Tuple2(0x1F210, 0x1F219),
-  Tuple2(0x1F21B, 0x1F22E),
-  Tuple2(0x1F230, 0x1F231),
-  Tuple2(0x1F23B, 0x1F23B),
+  Tuple2(0x1F004, 0x1F004),
+  Tuple2(0x1F0CF, 0x1F0CF),
+  Tuple2(0x1F18E, 0x1F18E),
+  Tuple2(0x1F191, 0x1F19A),
+  Tuple2(0x1F200, 0x1F202),
+  Tuple2(0x1F210, 0x1F23B),
   Tuple2(0x1F240, 0x1F248),
+  Tuple2(0x1F250, 0x1F251),
   Tuple2(0x1F260, 0x1F265),
+  Tuple2(0x1F300, 0x1F320),
+  Tuple2(0x1F32D, 0x1F335),
+  Tuple2(0x1F337, 0x1F37C),
+  Tuple2(0x1F37E, 0x1F393),
+  Tuple2(0x1F3A0, 0x1F3CA),
+  Tuple2(0x1F3CF, 0x1F3D3),
+  Tuple2(0x1F3E0, 0x1F3F0),
+  Tuple2(0x1F3F4, 0x1F3F4),
+  Tuple2(0x1F3F8, 0x1F43E),
+  Tuple2(0x1F440, 0x1F440),
+  Tuple2(0x1F442, 0x1F4FC),
+  Tuple2(0x1F4FF, 0x1F53D),
+  Tuple2(0x1F54B, 0x1F54E),
+  Tuple2(0x1F550, 0x1F567),
+  Tuple2(0x1F595, 0x1F596),
+  Tuple2(0x1F5FB, 0x1F64F),
+  Tuple2(0x1F680, 0x1F6C5),
+  Tuple2(0x1F6CC, 0x1F6CC),
+  Tuple2(0x1F6D0, 0x1F6D0),
   Tuple2(0x1F6D5, 0x1F6D7),
+  Tuple2(0x1F6DD, 0x1F6DF),
+  Tuple2(0x1F6EB, 0x1F6EC),
   Tuple2(0x1F6F9, 0x1F6FC),
   Tuple2(0x1F7E0, 0x1F7EB),
-  Tuple2(0x1F90C, 0x1F90F),
+  Tuple2(0x1F7F0, 0x1F7F0),
+  Tuple2(0x1F90C, 0x1F918),
   Tuple2(0x1F93F, 0x1F93F),
   Tuple2(0x1F94D, 0x1F94F),
-  Tuple2(0x1F96C, 0x1F978),
-  Tuple2(0x1F97A, 0x1F97F),
-  Tuple2(0x1F998, 0x1F9BF),
-  Tuple2(0x1F9C1, 0x1F9CB),
-  Tuple2(0x1F9CD, 0x1F9CF),
+  Tuple2(0x1F96C, 0x1F984),
+  Tuple2(0x1F998, 0x1F9CF),
   Tuple2(0x1F9E7, 0x1F9FF),
   Tuple2(0x1FA70, 0x1FA74),
-  Tuple2(0x1FA78, 0x1FA7A),
+  Tuple2(0x1FA78, 0x1FA7C),
   Tuple2(0x1FA80, 0x1FA86),
-  Tuple2(0x1FA90, 0x1FAA8),
-  Tuple2(0x1FAB0, 0x1FAB6),
-  Tuple2(0x1FAC0, 0x1FAC2),
-  Tuple2(0x1FAD0, 0x1FAD6),
+  Tuple2(0x1FA90, 0x1FAAC),
+  Tuple2(0x1FAB0, 0x1FABA),
+  Tuple2(0x1FAC0, 0x1FAC5),
+  Tuple2(0x1FAD0, 0x1FAD9),
+  Tuple2(0x1FAE0, 0x1FAE7),
+  Tuple2(0x1FAF0, 0x1FAF6),
   Tuple2(0x20000, 0x2FFFD),
   Tuple2(0x30000, 0x3FFFD),
 ];
@@ -442,7 +478,6 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x5C8, 0x5CF),
   Tuple2(0x5EB, 0x5EE),
   Tuple2(0x5F5, 0x5FF),
-  Tuple2(0x61D, 0x61D),
   Tuple2(0x70E, 0x70E),
   Tuple2(0x74B, 0x74C),
   Tuple2(0x7B2, 0x7BF),
@@ -451,9 +486,9 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x83F, 0x83F),
   Tuple2(0x85C, 0x85D),
   Tuple2(0x85F, 0x85F),
-  Tuple2(0x86B, 0x89F),
-  Tuple2(0x8B5, 0x8B5),
-  Tuple2(0x8C8, 0x8D2),
+  Tuple2(0x86B, 0x86F),
+  Tuple2(0x88F, 0x88F),
+  Tuple2(0x892, 0x897),
   Tuple2(0x984, 0x984),
   Tuple2(0x98D, 0x98E),
   Tuple2(0x991, 0x992),
@@ -531,12 +566,13 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0xC0D, 0xC0D),
   Tuple2(0xC11, 0xC11),
   Tuple2(0xC29, 0xC29),
-  Tuple2(0xC3A, 0xC3C),
+  Tuple2(0xC3A, 0xC3B),
   Tuple2(0xC45, 0xC45),
   Tuple2(0xC49, 0xC49),
   Tuple2(0xC4E, 0xC54),
   Tuple2(0xC57, 0xC57),
-  Tuple2(0xC5B, 0xC5F),
+  Tuple2(0xC5B, 0xC5C),
+  Tuple2(0xC5E, 0xC5F),
   Tuple2(0xC64, 0xC65),
   Tuple2(0xC70, 0xC76),
   Tuple2(0xC8D, 0xC8D),
@@ -547,7 +583,7 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0xCC5, 0xCC5),
   Tuple2(0xCC9, 0xCC9),
   Tuple2(0xCCE, 0xCD4),
-  Tuple2(0xCD7, 0xCDD),
+  Tuple2(0xCD7, 0xCDC),
   Tuple2(0xCDF, 0xCDF),
   Tuple2(0xCE4, 0xCE5),
   Tuple2(0xCF0, 0xCF0),
@@ -615,8 +651,7 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x13FE, 0x13FF),
   Tuple2(0x169D, 0x169F),
   Tuple2(0x16F9, 0x16FF),
-  Tuple2(0x170D, 0x170D),
-  Tuple2(0x1715, 0x171F),
+  Tuple2(0x1716, 0x171E),
   Tuple2(0x1737, 0x173F),
   Tuple2(0x1754, 0x175F),
   Tuple2(0x176D, 0x176D),
@@ -625,7 +660,6 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x17DE, 0x17DF),
   Tuple2(0x17EA, 0x17EF),
   Tuple2(0x17FA, 0x17FF),
-  Tuple2(0x180F, 0x180F),
   Tuple2(0x181A, 0x181F),
   Tuple2(0x1879, 0x187F),
   Tuple2(0x18AB, 0x18AF),
@@ -645,9 +679,9 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x1A8A, 0x1A8F),
   Tuple2(0x1A9A, 0x1A9F),
   Tuple2(0x1AAE, 0x1AAF),
-  Tuple2(0x1AC1, 0x1AFF),
-  Tuple2(0x1B4C, 0x1B4F),
-  Tuple2(0x1B7D, 0x1B7F),
+  Tuple2(0x1ACF, 0x1AFF),
+  Tuple2(0x1B4D, 0x1B4F),
+  Tuple2(0x1B7F, 0x1B7F),
   Tuple2(0x1BF4, 0x1BFB),
   Tuple2(0x1C38, 0x1C3A),
   Tuple2(0x1C4A, 0x1C4C),
@@ -655,7 +689,6 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x1CBB, 0x1CBC),
   Tuple2(0x1CC8, 0x1CCF),
   Tuple2(0x1CFB, 0x1CFF),
-  Tuple2(0x1DFA, 0x1DFA),
   Tuple2(0x1F16, 0x1F17),
   Tuple2(0x1F1E, 0x1F1F),
   Tuple2(0x1F46, 0x1F47),
@@ -676,15 +709,13 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x2072, 0x2073),
   Tuple2(0x208F, 0x208F),
   Tuple2(0x209D, 0x209F),
-  Tuple2(0x20C0, 0x20CF),
+  Tuple2(0x20C1, 0x20CF),
   Tuple2(0x20F1, 0x20FF),
   Tuple2(0x218C, 0x218F),
   Tuple2(0x2427, 0x243F),
   Tuple2(0x244B, 0x245F),
   Tuple2(0x2B74, 0x2B75),
   Tuple2(0x2B96, 0x2B96),
-  Tuple2(0x2C2F, 0x2C2F),
-  Tuple2(0x2C5F, 0x2C5F),
   Tuple2(0x2CF4, 0x2CF8),
   Tuple2(0x2D26, 0x2D26),
   Tuple2(0x2D28, 0x2D2C),
@@ -700,7 +731,7 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x2DCF, 0x2DCF),
   Tuple2(0x2DD7, 0x2DD7),
   Tuple2(0x2DDF, 0x2DDF),
-  Tuple2(0x2E53, 0x2E7F),
+  Tuple2(0x2E5E, 0x2E7F),
   Tuple2(0x2E9A, 0x2E9A),
   Tuple2(0x2EF4, 0x2EFF),
   Tuple2(0x2FD6, 0x2FEF),
@@ -713,14 +744,15 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x31E4, 0x31EF),
   Tuple2(0x321F, 0x321F),
   Tuple2(0x3401, 0x4DBE),
-  Tuple2(0x4E01, 0x9FFB),
-  Tuple2(0x9FFD, 0x9FFF),
+  Tuple2(0x4E01, 0x9FFE),
   Tuple2(0xA48D, 0xA48F),
   Tuple2(0xA4C7, 0xA4CF),
   Tuple2(0xA62C, 0xA63F),
   Tuple2(0xA6F8, 0xA6FF),
-  Tuple2(0xA7C0, 0xA7C1),
-  Tuple2(0xA7CB, 0xA7F4),
+  Tuple2(0xA7CB, 0xA7CF),
+  Tuple2(0xA7D2, 0xA7D2),
+  Tuple2(0xA7D4, 0xA7D4),
+  Tuple2(0xA7DA, 0xA7F1),
   Tuple2(0xA82D, 0xA82F),
   Tuple2(0xA83A, 0xA83F),
   Tuple2(0xA878, 0xA87F),
@@ -757,11 +789,10 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0xFB3F, 0xFB3F),
   Tuple2(0xFB42, 0xFB42),
   Tuple2(0xFB45, 0xFB45),
-  Tuple2(0xFBC2, 0xFBD2),
-  Tuple2(0xFD40, 0xFD4F),
+  Tuple2(0xFBC3, 0xFBD2),
   Tuple2(0xFD90, 0xFD91),
-  Tuple2(0xFDC8, 0xFDEF),
-  Tuple2(0xFDFE, 0xFDFF),
+  Tuple2(0xFDC8, 0xFDCE),
+  Tuple2(0xFDD0, 0xFDEF),
   Tuple2(0xFE1A, 0xFE1F),
   Tuple2(0xFE53, 0xFE53),
   Tuple2(0xFE67, 0xFE67),
@@ -805,10 +836,20 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x104FC, 0x104FF),
   Tuple2(0x10528, 0x1052F),
   Tuple2(0x10564, 0x1056E),
-  Tuple2(0x10570, 0x105FF),
+  Tuple2(0x1057B, 0x1057B),
+  Tuple2(0x1058B, 0x1058B),
+  Tuple2(0x10593, 0x10593),
+  Tuple2(0x10596, 0x10596),
+  Tuple2(0x105A2, 0x105A2),
+  Tuple2(0x105B2, 0x105B2),
+  Tuple2(0x105BA, 0x105BA),
+  Tuple2(0x105BD, 0x105FF),
   Tuple2(0x10737, 0x1073F),
   Tuple2(0x10756, 0x1075F),
-  Tuple2(0x10768, 0x107FF),
+  Tuple2(0x10768, 0x1077F),
+  Tuple2(0x10786, 0x10786),
+  Tuple2(0x107B1, 0x107B1),
+  Tuple2(0x107BB, 0x107FF),
   Tuple2(0x10806, 0x10807),
   Tuple2(0x10809, 0x10809),
   Tuple2(0x10836, 0x10836),
@@ -851,12 +892,13 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x10EAE, 0x10EAF),
   Tuple2(0x10EB2, 0x10EFF),
   Tuple2(0x10F28, 0x10F2F),
-  Tuple2(0x10F5A, 0x10FAF),
+  Tuple2(0x10F5A, 0x10F6F),
+  Tuple2(0x10F8A, 0x10FAF),
   Tuple2(0x10FCC, 0x10FDF),
   Tuple2(0x10FF7, 0x10FFF),
   Tuple2(0x1104E, 0x11051),
-  Tuple2(0x11070, 0x1107E),
-  Tuple2(0x110C2, 0x110CC),
+  Tuple2(0x11076, 0x1107E),
+  Tuple2(0x110C3, 0x110CC),
   Tuple2(0x110CE, 0x110CF),
   Tuple2(0x110E9, 0x110EF),
   Tuple2(0x110FA, 0x110FF),
@@ -898,11 +940,11 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x11645, 0x1164F),
   Tuple2(0x1165A, 0x1165F),
   Tuple2(0x1166D, 0x1167F),
-  Tuple2(0x116B9, 0x116BF),
+  Tuple2(0x116BA, 0x116BF),
   Tuple2(0x116CA, 0x116FF),
   Tuple2(0x1171B, 0x1171C),
   Tuple2(0x1172C, 0x1172F),
-  Tuple2(0x11740, 0x117FF),
+  Tuple2(0x11747, 0x117FF),
   Tuple2(0x1183C, 0x1189F),
   Tuple2(0x118F3, 0x118FE),
   Tuple2(0x11907, 0x11908),
@@ -917,7 +959,7 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x119D8, 0x119D9),
   Tuple2(0x119E5, 0x119FF),
   Tuple2(0x11A48, 0x11A4F),
-  Tuple2(0x11AA3, 0x11ABF),
+  Tuple2(0x11AA3, 0x11AAF),
   Tuple2(0x11AF9, 0x11BFF),
   Tuple2(0x11C09, 0x11C09),
   Tuple2(0x11C37, 0x11C37),
@@ -945,14 +987,16 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x1239A, 0x123FF),
   Tuple2(0x1246F, 0x1246F),
   Tuple2(0x12475, 0x1247F),
-  Tuple2(0x12544, 0x12FFF),
+  Tuple2(0x12544, 0x12F8F),
+  Tuple2(0x12FF3, 0x12FFF),
   Tuple2(0x1342F, 0x1342F),
   Tuple2(0x13439, 0x143FF),
   Tuple2(0x14647, 0x167FF),
   Tuple2(0x16A39, 0x16A3F),
   Tuple2(0x16A5F, 0x16A5F),
   Tuple2(0x16A6A, 0x16A6D),
-  Tuple2(0x16A70, 0x16ACF),
+  Tuple2(0x16ABF, 0x16ABF),
+  Tuple2(0x16ACA, 0x16ACF),
   Tuple2(0x16AEE, 0x16AEF),
   Tuple2(0x16AF6, 0x16AFF),
   Tuple2(0x16B46, 0x16B4F),
@@ -970,8 +1014,11 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x187F8, 0x187FF),
   Tuple2(0x18CD6, 0x18CFF),
   Tuple2(0x18D01, 0x18D07),
-  Tuple2(0x18D09, 0x1AFFF),
-  Tuple2(0x1B11F, 0x1B14F),
+  Tuple2(0x18D09, 0x1AFEF),
+  Tuple2(0x1AFF4, 0x1AFF4),
+  Tuple2(0x1AFFC, 0x1AFFC),
+  Tuple2(0x1AFFF, 0x1AFFF),
+  Tuple2(0x1B123, 0x1B14F),
   Tuple2(0x1B153, 0x1B163),
   Tuple2(0x1B168, 0x1B16F),
   Tuple2(0x1B2FC, 0x1BBFF),
@@ -979,10 +1026,13 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x1BC7D, 0x1BC7F),
   Tuple2(0x1BC89, 0x1BC8F),
   Tuple2(0x1BC9A, 0x1BC9B),
-  Tuple2(0x1BCA4, 0x1CFFF),
+  Tuple2(0x1BCA4, 0x1CEFF),
+  Tuple2(0x1CF2E, 0x1CF2F),
+  Tuple2(0x1CF47, 0x1CF4F),
+  Tuple2(0x1CFC4, 0x1CFFF),
   Tuple2(0x1D0F6, 0x1D0FF),
   Tuple2(0x1D127, 0x1D128),
-  Tuple2(0x1D1E9, 0x1D1FF),
+  Tuple2(0x1D1EB, 0x1D1FF),
   Tuple2(0x1D246, 0x1D2DF),
   Tuple2(0x1D2F4, 0x1D2FF),
   Tuple2(0x1D357, 0x1D35F),
@@ -1009,7 +1059,8 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x1D7CC, 0x1D7CD),
   Tuple2(0x1DA8C, 0x1DA9A),
   Tuple2(0x1DAA0, 0x1DAA0),
-  Tuple2(0x1DAB0, 0x1DFFF),
+  Tuple2(0x1DAB0, 0x1DEFF),
+  Tuple2(0x1DF1F, 0x1DFFF),
   Tuple2(0x1E007, 0x1E007),
   Tuple2(0x1E019, 0x1E01A),
   Tuple2(0x1E022, 0x1E022),
@@ -1018,9 +1069,14 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x1E12D, 0x1E12F),
   Tuple2(0x1E13E, 0x1E13F),
   Tuple2(0x1E14A, 0x1E14D),
-  Tuple2(0x1E150, 0x1E2BF),
+  Tuple2(0x1E150, 0x1E28F),
+  Tuple2(0x1E2AF, 0x1E2BF),
   Tuple2(0x1E2FA, 0x1E2FE),
-  Tuple2(0x1E300, 0x1E7FF),
+  Tuple2(0x1E300, 0x1E7DF),
+  Tuple2(0x1E7E7, 0x1E7E7),
+  Tuple2(0x1E7EC, 0x1E7EC),
+  Tuple2(0x1E7EF, 0x1E7EF),
+  Tuple2(0x1E7FF, 0x1E7FF),
   Tuple2(0x1E8C5, 0x1E8C6),
   Tuple2(0x1E8D7, 0x1E8FF),
   Tuple2(0x1E94C, 0x1E94F),
@@ -1074,36 +1130,37 @@ List<Tuple2> _unassigned_table = [
   Tuple2(0x1F249, 0x1F24F),
   Tuple2(0x1F252, 0x1F25F),
   Tuple2(0x1F266, 0x1F2FF),
-  Tuple2(0x1F6D8, 0x1F6DF),
+  Tuple2(0x1F6D8, 0x1F6DC),
   Tuple2(0x1F6ED, 0x1F6EF),
   Tuple2(0x1F6FD, 0x1F6FF),
   Tuple2(0x1F774, 0x1F77F),
   Tuple2(0x1F7D9, 0x1F7DF),
-  Tuple2(0x1F7EC, 0x1F7FF),
+  Tuple2(0x1F7EC, 0x1F7EF),
+  Tuple2(0x1F7F1, 0x1F7FF),
   Tuple2(0x1F80C, 0x1F80F),
   Tuple2(0x1F848, 0x1F84F),
   Tuple2(0x1F85A, 0x1F85F),
   Tuple2(0x1F888, 0x1F88F),
   Tuple2(0x1F8AE, 0x1F8AF),
   Tuple2(0x1F8B2, 0x1F8FF),
-  Tuple2(0x1F979, 0x1F979),
-  Tuple2(0x1F9CC, 0x1F9CC),
   Tuple2(0x1FA54, 0x1FA5F),
   Tuple2(0x1FA6E, 0x1FA6F),
   Tuple2(0x1FA75, 0x1FA77),
-  Tuple2(0x1FA7B, 0x1FA7F),
+  Tuple2(0x1FA7D, 0x1FA7F),
   Tuple2(0x1FA87, 0x1FA8F),
-  Tuple2(0x1FAA9, 0x1FAAF),
-  Tuple2(0x1FAB7, 0x1FABF),
-  Tuple2(0x1FAC3, 0x1FACF),
-  Tuple2(0x1FAD7, 0x1FAFF),
+  Tuple2(0x1FAAD, 0x1FAAF),
+  Tuple2(0x1FABB, 0x1FABF),
+  Tuple2(0x1FAC6, 0x1FACF),
+  Tuple2(0x1FADA, 0x1FADF),
+  Tuple2(0x1FAE8, 0x1FAEF),
+  Tuple2(0x1FAF7, 0x1FAFF),
   Tuple2(0x1FB93, 0x1FB93),
   Tuple2(0x1FBCB, 0x1FBEF),
   Tuple2(0x1FBFA, 0x1FFFF),
-  Tuple2(0x20001, 0x2A6DC),
-  Tuple2(0x2A6DE, 0x2A6FF),
-  Tuple2(0x2A701, 0x2B733),
-  Tuple2(0x2B735, 0x2B73F),
+  Tuple2(0x20001, 0x2A6DE),
+  Tuple2(0x2A6E0, 0x2A6FF),
+  Tuple2(0x2A701, 0x2B737),
+  Tuple2(0x2B739, 0x2B73F),
   Tuple2(0x2B741, 0x2B81C),
   Tuple2(0x2B81E, 0x2B81F),
   Tuple2(0x2B821, 0x2CEA0),
@@ -1294,8 +1351,7 @@ List<Tuple2> _ambiguous_table = [
   Tuple2(0x1F100, 0x1F10A),
   Tuple2(0x1F110, 0x1F12D),
   Tuple2(0x1F130, 0x1F169),
-  Tuple2(0x1F172, 0x1F17D),
-  Tuple2(0x1F180, 0x1F18D),
+  Tuple2(0x1F170, 0x1F18D),
   Tuple2(0x1F18F, 0x1F190),
   Tuple2(0x1F19B, 0x1F1AC),
   Tuple2(0xE0100, 0xE01EF),
@@ -1304,46 +1360,15 @@ List<Tuple2> _ambiguous_table = [
 ];
 
 List<Tuple2> _widened_table = [
-  Tuple2(0x1F004, 0x1F004),
-  Tuple2(0x1F0CF, 0x1F0CF),
-  Tuple2(0x1F170, 0x1F171),
-  Tuple2(0x1F17E, 0x1F17F),
-  Tuple2(0x1F18E, 0x1F18E),
-  Tuple2(0x1F191, 0x1F19A),
-  Tuple2(0x1F201, 0x1F202),
-  Tuple2(0x1F21A, 0x1F21A),
-  Tuple2(0x1F22F, 0x1F22F),
-  Tuple2(0x1F232, 0x1F23A),
-  Tuple2(0x1F250, 0x1F251),
-  Tuple2(0x1F300, 0x1F320),
-  Tuple2(0x1F32D, 0x1F335),
-  Tuple2(0x1F337, 0x1F37C),
-  Tuple2(0x1F37E, 0x1F393),
-  Tuple2(0x1F3A0, 0x1F3CA),
-  Tuple2(0x1F3CF, 0x1F3D3),
-  Tuple2(0x1F3E0, 0x1F3F0),
-  Tuple2(0x1F3F4, 0x1F3F4),
-  Tuple2(0x1F3F8, 0x1F43E),
-  Tuple2(0x1F440, 0x1F440),
-  Tuple2(0x1F442, 0x1F4FC),
-  Tuple2(0x1F4FF, 0x1F53D),
-  Tuple2(0x1F54B, 0x1F54E),
-  Tuple2(0x1F550, 0x1F567),
   Tuple2(0x1F57A, 0x1F57A),
-  Tuple2(0x1F595, 0x1F596),
   Tuple2(0x1F5A4, 0x1F5A4),
-  Tuple2(0x1F5FB, 0x1F64F),
-  Tuple2(0x1F680, 0x1F6C5),
-  Tuple2(0x1F6CC, 0x1F6CC),
-  Tuple2(0x1F6D0, 0x1F6D2),
-  Tuple2(0x1F6EB, 0x1F6EC),
+  Tuple2(0x1F6D1, 0x1F6D2),
   Tuple2(0x1F6F4, 0x1F6F8),
-  Tuple2(0x1F910, 0x1F93A),
+  Tuple2(0x1F919, 0x1F93A),
   Tuple2(0x1F93C, 0x1F93E),
   Tuple2(0x1F940, 0x1F945),
   Tuple2(0x1F947, 0x1F94C),
   Tuple2(0x1F950, 0x1F96B),
-  Tuple2(0x1F980, 0x1F997),
-  Tuple2(0x1F9C0, 0x1F9C0),
+  Tuple2(0x1F985, 0x1F997),
   Tuple2(0x1F9D0, 0x1F9E6),
 ];
